QA-Studio/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.6
python-socketio>=5.10.0
eventlet>=0.33.3
playwright>=1.40.0
pytest>=7.4.3
pytest-playwright>=0.4.3
pytest-xdist>=3.5.0
pydantic>=2.9.0
beautifulsoup4>=4.12.2
advertools>=0.14.3
Pillow>=10.2.0
PyYAML>=6.0.1
requests>=2.31.0
